chore: rename correctly internal folders
